TF-A: Fix non-standard frequency issue in udelay
Fixes a non-standard frequency issue, which made impossible usage of
99.99MHz frequency and alike.

+++ include/drivers/delay_timer.h
@@ -22,7 +22,7 @@
  ********************************************************************/
 
 typedef struct timer_ops {
-	uint32_t (*get_timer_value)(void);
+	uint64_t (*get_timer_value)(void);
 	uint32_t clk_mult;
 	uint32_t clk_div;
 } timer_ops_t;
